* V5-7-patches:
  Fix agent crash in expression mib when syntax error in expr or inacessible variables.

+++ agent/mibgroup/disman/expr/expValue.c
@@ -731,7 +731,7 @@ expValue_evaluateExpression( struct expExpression *exp,
     char exprRaw[     EXP_STR3_LEN+1 ];
     netsnmp_variable_list *var;
     netsnmp_variable_list owner_var, name_var, param_var;
-    long n;
+    int n;
     char *cp;
 
     if (!exp)
